+++ src/Tests/Formatting.Analyzers.Tests/RCS0063RemoveUnnecessaryBlankLineTests.cs
@@ -2,17 +2,17 @@
 
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Roslynator.CSharp.CodeFixes;
+using Roslynator.Formatting.CodeFixes.CSharp;
 using Roslynator.Testing.CSharp;
 using Xunit;
 
-namespace Roslynator.CSharp.Analysis.Tests;
+namespace Roslynator.Formatting.CSharp.Tests;
 
-public class RCS1036RemoveUnnecessaryBlankLineTests : AbstractCSharpDiagnosticVerifier<RemoveUnnecessaryBlankLineAnalyzer, WhitespaceTriviaCodeFixProvider>
+public class RCS0063RemoveUnnecessaryBlankLineTests : AbstractCSharpDiagnosticVerifier<RemoveUnnecessaryBlankLineAnalyzer, RemoveUnnecessaryBlankLineFixProvider>
 {
-    public override DiagnosticDescriptor Descriptor { get; } = DiagnosticRules.Obsolete_RemoveUnnecessaryBlankLine;
+    public override DiagnosticDescriptor Descriptor { get; } = DiagnosticRules.RemoveUnnecessaryBlankLine;
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_ObjectInitializer()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -50,7 +50,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_ObjectInitializer_WithTrailingComma()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -88,7 +88,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_ArrayInitializer()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -120,7 +120,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_CollectionInitializer()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -156,7 +156,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyDeclaration()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -171,7 +171,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyBlock()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -192,7 +192,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyLineAfterDocComment()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -215,7 +215,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EndOfFile()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -230,7 +230,7 @@ class C
 ", options: Options.EnableConfigOption(ConfigOptionKeys.BlankLineBetweenClosingBraceAndSwitchSection));
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EndOfFile2()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -246,7 +246,7 @@ class C
 ", options: Options.EnableConfigOption(ConfigOptionKeys.BlankLineBetweenClosingBraceAndSwitchSection));
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_LastEmptyLineInDoStatement()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -275,7 +275,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyLineAfterLastEnumMember_NoTrailingComma()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -300,7 +300,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyLineAfterLastEnumMember_TrailingComma()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -325,7 +325,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyLineBeforeFirstEnumMember()
     {
         await VerifyDiagnosticAndFixAsync(@"
@@ -350,7 +350,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_ObjectInitializer()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -371,7 +371,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_ObjectInitializer_Singleline()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -388,7 +388,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_ObjectInitializer_Empty()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -402,7 +402,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineBetweenClosingBraceAndSwitchSection()
     {
         await VerifyNoDiagnosticAsync("""
@@ -428,7 +428,7 @@ class C
 """);
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineAtEndOfFile()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -438,7 +438,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineAtEndOfFileAfterMultiLineComment()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -449,7 +449,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineAtEndOfFileWithWhitespace()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -459,7 +459,7 @@ class C
  ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineAtEndOfFileAfterSingleLineComment()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -470,7 +470,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task TestNoDiagnostic_EmptyLineAtEndOfFileAfterPreprocessorDirective()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -482,7 +482,7 @@ class C
 ");
     }
 
-    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.Obsolete_RemoveUnnecessaryBlankLine)]
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBlankLine)]
     public async Task Test_EmptyLineBetweenClosingBraceAndSwitchSection()
     {
         await VerifyNoDiagnosticAsync("""
