+++ src/VisualStudioCode/package/src/configurationFiles.generated.ts
@@ -51,7 +51,7 @@ roslynator_analyzers.enabled_by_default = true|false
 # Applicable to: rcs0060
 
 #roslynator_blank_line_between_closing_brace_and_switch_section = true|false
-# Applicable to: rcs0014, rcs1036
+# Applicable to: rcs0014, rcs0063, rcs1036
 
 #roslynator_blank_line_between_single_line_accessors = true|false
 # Applicable to: rcs0011
@@ -309,6 +309,10 @@ roslynator_analyzers.enabled_by_default = true|false
 #dotnet_diagnostic.rcs0061.severity = none
 # Options: roslynator_blank_line_between_switch_sections
 
+# Remove unnecessary blank line
+#dotnet_diagnostic.rcs0063.severity = none
+# Options: roslynator_blank_line_between_closing_brace_and_switch_section
+
 # Add braces (when expression spans over multiple lines)
 #dotnet_diagnostic.rcs1001.severity = suggestion
 
@@ -368,10 +372,6 @@ roslynator_analyzers.enabled_by_default = true|false
 
 # Remove redundant 'sealed' modifier
 #dotnet_diagnostic.rcs1034.severity = silent
-
-# Remove unnecessary blank line
-#dotnet_diagnostic.rcs1036.severity = suggestion
-# Options: roslynator_blank_line_between_closing_brace_and_switch_section
 
 # Remove trailing white-space
 #dotnet_diagnostic.rcs1037.severity = suggestion
